+++ server/db/models/ledger.js
@@ -4,7 +4,7 @@ const db = require('../db')
 const Ledger = db.define('ledger', {
   value: {type: Sequelize.INTEGER, allowNull: false},
   shares: {type: Sequelize.INTEGER, allowNull: true},
-  units: {type: Sequelize.DECIMAL, allowNull: true}
+  units: {type: Sequelize.INTEGER, allowNull: true}
 })
 
 module.exports = Ledger
